Answer for Ch. 5, Ex 2 of Learning Perl Student Workbook
Modify your answer for Exercise 5.1 so after it reads lines from the
files you specify on the command line, it reads in lines from standard
input and prefaces the line with stdin.

I ended up doing this quite differently than the workbook. The workbook
says to open '-' as the file, but my particular version of perl didn't
allow that. Not sure if that's an OS X thing, or a perl version thing.

/rs.