Answer for Ch. 5, Ex 7 of Learning Perl Student Workbook
Write a program that goes through the list of files from
the command line and reports their line count. Use warn to
print a message for each file that does not exist.

I did this differently that the book, in that I a) checked if the file
couldn't be opened first, rather than last, and b) didn't use a
subroutine to read all of the lines. I didn't consider using one, since
it was such a short problem.

/rs.