Answer for Ch. 5, Ex 13 of Learning Perl Student Workbook
Write a program that converts a file encoded as UTF-8 into UTF-16BE.
Use Perl’s encoding layers to read the data in the correct format
and output the same data in the new encoding.

Again, trivially simple to add another filename, and change the
encodings for the input and output files.

/rs.