Answer for Ch. 6, Ex 2 of Learning Perl
Write a program that reads a series of words (with one word per
line) until end-of-input, then prints a summary of how many
times each word was seen.

Read the list of words into an array, count each word, storing the
number in the hash, keyed by word, then print each key (sorted by word).

I originally had the first foreach loop checking to see if the word key
was defined in the hash, but changed it after being reminded by the text
that if it's undefined, perl will happily convert undef to a 0 since
we're treating it as a number. Shortcuts FTW!

/rs.