Answer for Ch. 6, Ex 3 of Learning Perl
Write a program to list all of the keys and values in %ENV. Print the
results in two columns in ASCIIbetical order. For extra credit, arrange
the output to vertically align both columns.

Simple pair of loops, first to read over the hash to find the longest
key, and the second to loop over the hash to print them.

I spent more time on making the pretty header / footer for the table,
with appropriate length lines and whatnot. It also looks best with a
wide terminal window.

/rs.