Answer for Ch. 6, Ex. 4 of Learning Perl Student Workbook.
What happens in your program for Exercise 6.3 if the user-entered name
is in the hash, but does not have a birthday? Add a couple of entries to
the hash: one with a value of undef, and one with a value of the empty
string. Modify your program for Exercise 6.3 to tell the user that you
don’t know the birthday if the key is in the hash, but does not have a
value.

This was a LITTLE more complex than I thought. I needed to check for the
undefined entries that still had keys in the hash, not just check for
undefined keys. Still pretty simple.

/rs.