Answer for Ch. 6, Ex. 7 of Learning Perl Student Workbook.
In Exercise 6.1, you wrote a program to report the entries in the hash,
and since then you’ve added keys that have values that are undefined or
the empty string. Modify that first program to delete from the hash the
keys that don’t have true values then print the contents of the hash.

I initially did all of this in one foreach loop, but changed it to match
the answer in the workbook, since it shows a little more explicitly
what's going on. I mean, the output is the same as the first exercise
either way.

/rs.