Answer for Ch. 7, Ex 1 of Learning Perl
Make a program that prints each line of its input that mentions fred.
(It shouldn’t do anything for other lines of input.) Does it match if
your input string is Fred, frederick, or Alfred? Make a small text file
with a few lines mentioning “fred flintstone” and his friends, then use
that file as input to this program and the ones later in this section.”

Pretty simple, just read in a file, go through it line by line, and test
if the line matches the regex /fred/. My answer was more complex than
the book's, with error checking, and explicitly opening the file.

I also added a test data file, 'testdata'.

/rs.