Answer for Ch. 7, Ex 5 of Learning Perl
Make a program that prints each line that has a two of the same
nonwhitespace characters next to each other. It should match lines that
contain words such as Mississippi, Bamm-Bamm, or llama.

I used Unicode Properties to match alpha characters, because those
weren't around when I started with perl, and I think they're hella neat.
I also used the /g notation for referring to backreferences. And it
worked!

/rs.