Answer for Ch. 7, Ex 6 of Learning Perl
Write a program that prints out any input line that mentions both wilma
and fred.

My answer is a little more complex than the book's, as I check for
Capitalized Names.

I also nearly made a mistake. I originally searched for a pattern of
Wilma or Fred with characters in between Wilma or Fred. This caught any
mention of both characters in a line, but it also caught the repeated
ine, which wasn't correct.

So I changed it to check for Wilma with characters between Fred or Fred
with characters between Wilma, and that works.

/rs.