Answer for Ch. 7, Ex. 1 of Learning Perl Student Workbook.
Write a program that matches a string containing either “fred” or
“barney” then reports the part of the string before the match, the part
of the string after the match, and the name that it matched.

My answer differs significantly from the workbook's. I used
backreference variables to pick out the parts I wanted, whereas the
answer uses the /p flag and PREMATCH/MATCH/POSTMATCH variables.

My question is, which is the better way to do this? Is there a
performance hit for using backreference variables? Is it worse than the
performance hit associated with the /p flag? These are the questions
that plague me at 4 in the morning, sick with the flu, unable to sleep.

/rs.