Answer for Ch. 7, Ex. 4 of Learning Perl Student Workbook.
Write a program to match a string that contains an “a”, but doesn’t have
a “b” anywhere after the “a”.

Initially, I tried this with a negated character class (i.e. [^b]), but
it appears that only works with shortcuts? I'll have to look it up to be
sure I didn't miss something.

Simplest way to do this was just a pair of nested ifs with super short
patterns to match.

/rs.