Answer for Ch. 7, Ex. 5 of Learning Perl Student Workbook.
Use a relative backreference to construct a pattern to match a sequence
of at least three characters that repeats itself at some point in the
string.

I went down a couple of different paths this time.

First, I matched any charater that repeats itself within the string
three times. A neat regex, but not correct. Would have lost marks on an
exam for this!

Second, I used the period to search for any character. This was great,
but lead to whitespace characters being found. This is technically
corret, but it didn't feel like it was really right.

Finally, I used the Unicode Property shortcut to match any three Alpha
characters, and that did what I felt it should do. I like Unicode
Properties!

/rs.