Answer for Ch. 7, Ex. 6 of Learning Perl Student Workbook.
Use relative backreferences to construct a pattern to match a sequence
of at least three characters that reports itself, but interwoven with
another sequence that repeats itself.

I added another group to match a three-pattern, and added a second
relative reference. I changed back to dots instead of unicode
properties, because it was shorter.

/rs.