Answer for Ch. 7, Ex. 7 of Learning Perl Student Workbook.
Use the Unicode properties to match lines that have numbers or
whitespace. If the line has a character with the number property, output
a message saying so. If the line has a character with whitespace, output
a message saying so. If it has both, output two messages.

Added a new test data file with random strings for this one. Pretty
simple solution, again.

/rs.