Answer for Ch. 8, Ex 2 of Learning Perl
Using the pattern test program, make a pattern that matches if any word
(in the \w sense of word) ends with the letter a. Does it match wilma
but not barney? Does it match Mrs. Wilma Flintstone? What about
wilma&fred? Try it on the sample text file from the Exercises in
Chapter 7.

It matches:
1) yes - wilma but not marney
2) yes - Mrs. Wilma Flintstone
3) yes - wilma&fred

The third matches because the word boundry isn't looking for spaces,
it's looking for the ends of words, and the ampersand denotes the end of
a word.

/rs.