Answer for Ch. 8, Ex 3 of Learning Perl
Modify the program from the previous exercise so that the word ending
with the letter a is captured into $1. Update the code to display that
variable’s contents in single quotes, something like $1 contains
'Wilma'.

Apparantly, I did most of this in the last exercise, because I didn't
need to actually change my regex pattern at all.

/rs.