Answer for Ch. 8, Ex 5 of Learning Perl
Modify the program from the previous exercise so that immediately
following the word ending in a it will also capture up-to-five
characters (if there are that many characters, of course) in a separate
capture variable. Update the code to display both capture variables.

The pattern still matches just plain 'wilma'.

/rs.