Answer for Ch. 8, Ex 6 of Learning Perl
Write a new program (not the test program!) that prints out any input
line ending with whitespace (other than just a newline). Put a marker
character at the end of the output line so as to make the whitespace
visible.

I interpreted this a little differently than the answer key. Instead of
just looking for a single whitespace character and printing the output
with a marker at the end, I captured all whitespace at the end of a
line, and printed the line out with markers around that whitespace.

I like my solution better.

/rs.