Answer for Ch. 9, Ex 2 of Learning Perl
Write a program that makes a modified copy of a text file. In the copy,
every string Fred (case-insensitive) should be replaced with Larry. (So,
Manfred Mann should become ManLarry Mann.) The input file name should be
given on the command line (don’t ask the user!), and the output filename
should be the corresponding file name ending with .out.

/rs.