Answer for Ch. 9, Ex. 4 of Learning Perl Student Workbook.
Read lines from standard input and print them in four versions: the
original string, all uppercase, all lowercase, and all lowercase except
for the first character in the string. Use the case-shifting operators
inside double-quoted strings.

My answer is a little more complex than the workbook's. I used a regular
expression to make sure I was only capitalizing the first letter of the
string, while leaving the capitalization intact in the rest. Looking at
it, it's sort of abusing the case-shifting operators for it, but it's
the best kind of correct... "technically correct".

Anyhow, I like my answer better.

/rs.