Answer for Ch. 9, Ex. 6 of Learning Perl Student Workbook.
Create a file of tab-separated values, including a line that has the
names of each column of values. Write a program that reads in the file
and produces a report for each line.

I figured this one out mostly solo. It took me a while to remember that
I could load the first line of the file WITHOUT using a while loop. I
used a foreach to display each field, but a while...each might have been
a little more readable.

/rs.