Answer for Ch. 10, Ex 1 of Learning Perl
Make a program that will repeatedly ask the user to guess a secret
number from 1 to 100 until the user guesses the secret number. Your
program should pick the number at random by using the magical formula
int(1 + rand 100).[285] When the user guesses wrong, the program should
respond, “Too high” or “Too low.” If the user enters the word quit or
exit, or if the user enters a blank line, the program should quit. Of
course, if the user guesses correctly, the program should quit then as
well!

This program reminded me of the dumb programs that I used to make for
the C=64 when I was a kid. Of course, back then, I didn't know about
random number generation, regular expressions, or branching if
statements. It was fun to write, though, even though it was so simple.

/rs.