Answer for Ch. 10, Ex 2 of Learning Perl
Modify the program from the previous exercise to print extra debugging
information as it goes along, such as the secret number it chose.

Still a nifty throwback program, but now I got to feel clever while
setting my debug variable. I used an if expression modifier, a logical
and (&&), and the conditional or (//).

/rs.