Answer for Ch. 10, Ex 3 of Learning Perl
Modify the program from Exercise 3 in Chapter 6 (the environment
lister) to print (undefined value) for environment variables without a
value.

I had an additional step in my original code, because I have a tendency
to assign to variables before I use them, but otherwise, it's pretty
much the same as the book's.

/rs.