Answer for Ch. 10, Ex. 3 of Learning Perl Student Workbook.
Write a program that reads lines of input and splits it into words. Use
a hash to keep a count of the number of times you’ve seen each word then
create a report of the count for each word.

Very similar to Exercise 06-03, except it was splitting words out of
lines, instead of just counting individual words.

/rs.