Answer for Ch. 10, Ex. 5 of Learning Perl Student Workbook.
Write a program to count the number of lines in a Perl program. Skip
blank lines and lines that only have a comment.

I stumbled on the fact that \z doesn't match end of line if there's a
newline after end of line. Once I figured that out, a well-placed chomp
made everything hunky-dory again.

/rs.