Answer for Ch. 10, Ex. 6 of Learning Perl Student Workbook.
Perl has a special token, __END__, which marks the end of a program when
it shows up on a line by itself. Modify your answer to Exercise 10.5 so
that you stop reading lines when you encounter that token.

I fiddled around with whether or not to count the __END__ line as a line
of code, ultimately deciding (correctly, I believe) to include it.

/rs.