Answer for Ch. 11, Ex 1 of Learning Perl
Install the Module::CoreList module from CPAN. Print a list of all of
the modules that came with Perl 5.14. To build a hash whose keys are the
names of the modules that came with a given version of perl, use this
line:

        my %modules = %{ $Module::CoreList::version{5.014} };

I forgot about 'join', which was an easy way to add a newline each key
of the hash.

/rs.