Answer for Ch. 11, Ex 2 of Learning Perl
Write a program using DateTime to compute the interval between now and
a date that you enter as the year, month, and day on the command line.

This was kinda fun. I ended up checking to make sure that the user
actually had command line arguments, and used from_epoch to get the
date, rather than localtime, after I read the documentation. I also
checked if the entered date is in the past or the future, and altered
the string that's output based on that.

/rs.