Answer for Ch. 11, Ex. 1 of Learning Perl Student Workbook.
Using the File::Spec module, write a program that takes a directory path
from the command line and adds it to every file in the current working
directory to create an absolute path.

The workbook can be a little frustrating, as their answers often use
language aspects that haven't been covered in the book. I've got
experience with the language, and am just using these to refresh, but a
true beginner would likely be lost.

Anyhow...

File::Spec->catfile is the key to this one. That's it.

/rs.