Answer for Ch. 11, Ex. 6 of Learning Perl Student Workbook.
Use the DBI module with the DBD::SQLite driver to create a small test
database for SQLite. From your Perl program, create a table named
“Characters” with columns for ID, First Name, and Last Name. Once you
create the table, insert the names of the Flintstones characters.

My SQL is rough, but I figured it out. I also took advantage of SQLite's
autoincrement, so I didn't have to keep track of it.

/rs.