Answer for Ch. 12, Ex. 1 of Learning Perl Student.
Make a program that takes a list of files named on the command line and
reports for each one whether it’s readable, writable, executable, or
doesn’t exist. (Hint: it may be helpful to have a function that will do
all of the file tests for one file at a time.) What does it report about
a file that has been chmod’ed to 0?  In most shells, use a star as
the argument to mean all of the normal files in the current directory.

My solution for this question is similar in basic function to the answer
in the text. I printed the status of the attribute regardless of whether
or not it was positive or negative, whereas the book's solution only
reported the positive attributes. The book also used an array for the
attributes, joining them into a string that is returned.

In retrospect, that solution is probably better than mine. Really, a
function should only return a value, and not output. While mine is
technically correct, it is not as clean or functional. It doesn't really
show that I've been picking up a good understanding of the language
either.

I'll need think a little more creatively in the future. It's okay to go
back and review and look up things if I'm not sure how to do it, and
it'll be a good reminder and learning experience if I do. Just remember
that.

/rs.