Answer for Ch. 12, Ex. 1 of Learning Perl Student.
Make a program to identify the oldest file named on the command line
and report its age in days. What does it do if the list is empty?

If a file isn't specified on the comman line, then the program complains
of errors trying to apply -M... that's why we check first to make sure
that something is passed to us.

I could have used list context to assign the temp values to the highest
values. I could also have calculated a proper time string from the days.

/rs.