Answer for Ch. 12, Ex. 2 of Learning Perl.
Make a program that uses stacked file test operators to list all files
named on the command line that are readable, writable, and owned by
you.

I used real owner, as opposed to effective owner, and I also tracked if
there were no files owned by the user and reported that.

Sometimes I do things differently and they're better, and they show an
understanding of things we've gone over, and sometimes, I do them
differently, and it's like I'm not even paying attention. This time, it
was the former.

/rs.