Answer for Ch. 12, Ex. 4 of Learning Perl Workbook.
Write a program that takes two numbers from the command line, prints
their binary representation then prints their logical-ANDed value.

This one has me confused. The bitwise-and works as expected when calling
it in the printf statement, but if I assigne the results of it to a
variable, it stops working, and gives me an incorrect answer. It looks
as if it's dropping trailing zeros from the binary number.

Gotta say, I'm confused about that one. I'll have to recall it for
later, and see if I can figure it out.

/rs.