Answer for Ch. 13, Ex. 1 of Learning Perl.
Write a program to ask the user for a directory name, then change to
that directory. If the user enters a line with nothing but whitespace,
change to his or her home directory as a default. After changing, list
the ordinary directory contents (not the items whose names begin with a
dot) in alphabetical order.

Pretty straightforward. I used unicode property classes, and found out
that using chdir without an argument is deprecated so I used an
environment variable to find the home directory.

/rs.