Answer for Ch. 13, Ex. 7 of Learning Perl.
If your operating system supports it, fix up the program from the
previous exercise to allow an optional -s switch before the other
arguments to indicate that you want to make a soft link instead of a
hard link.

I used Getopt::Std to parse arguments, because that's what you're
supposed to do, right? Use the modules available to you?

Aside from that, it's pretty much the same as the past three, aside from
a decision branch at the if.

/rs.