Answer for Ch. 13, Ex. 8 of Learning Perl.
If your operating system supports it, write a program to find any
symbolic links in the current directory and print out their values (like
ls -l would: name -> value).

I used a dirhandle, instead of a glob, and I didn't try to readlink
every file in the directory, only the ones that passed the -l test. I
don't know if my solution woul be more efficient than the text's or not!

/rs.