Answer for Ch. 13, Ex. 1 of Learning Perl Workbook.
Write a program that reports the number of files in the current
directory. Don’t count special files . or .., but count the “hidden”
files.

I could have used a glob, reading it into an array and getting the total
number of entries in it, too. Gotta remember to close my dirhandles when
I'm done with them, too!

/rs.