Answer for Ch. 13, Ex. 3 of Learning Perl Workbook.
Write a program that looks in the current working directory and creates
a col- umnar report of all files, their sizes, and whether they are
readable, writeable, and executable. Use opendir and a while loop.

I thought about using a conditional to assign the R/W/X values, but then
just defined them earlier on. It would probably be more efficient to use
the conditional, but I didn't feel like changing it.

/rs.