Answer for Ch. 13, Ex. 5 of Learning Perl Workbook.
Write a program that looks in the current working directory then lists
all of the Perl programs and their size. Use opendir to find the files.

I decided to make the assumption that perl programs ended in .pl. It's
naive, and a better solution would be to open each file, look at the
first line, and see if it's got a #!/ line with perl in it. That could
get complicated, though, if the file is a binary file and whatnot.

/rs.