Answer for Ch. 13, Ex. 7 of Learning Perl Workbook.
Use the utime function to set the access time of the files on the
command line to five minutes ago. Modify the program to set the
modification time of the files to the current system time.

This one was simpler than it seemed. The utime section was copying from
the book with slight modifications, and the output (that wasn't even
needed) was modified from a chapter 12 exercise.

/rs.