Answer for Ch. 14, Ex. 2 of Learning Perl.
Make a program that will print the last_name hash’s data sorted in
case-insensitive alphabetical order by last name. When the last names
are the same, sort those by first name (again, without regard for case).
That is, the first name in the output should be Fred’s, while the last
one should be Betty’s. All of the people with the same family name should
be grouped together. Don’t alter the data. The names should be printed
with the same capitalization.

Simple little cmp sort routine.

/rs.