Answer for Ch. 14, Ex. 1 of Learning Perl Workbook.
Write a program to read lines of input and report the position of the
first mention of “perl” on each line.

I could have read the input in the while loop, or even used the diamond
operator to read from a text file argument. I like prompting for it at
every line, though. UX, blah blah blah.

/rs.