Answer for Ch. 14, Ex. 3 of Learning Perl Workbook.
Write a program that takes a starting column value, an ending column, and a
filename. Go through the lines of the file and print only the portions of
each line from the start column to the end column. Use substr to get the
right part of the string.

Forgot that substr takes a length, not an ending column... I probably
would have figured it out if I had read the entire specification in the
book when I was reviewing. Oops.

/rs.