Answer for Ch. 14, Ex. 4 of Learning Perl Workbook.
Write a program to print a list of the files in the current directory,
sorted in order of their ascending file size. Change the program to make
it list the files in order of descending file size.

I needed to remember to add the 'keys' keyword to the sort by size line.
Made for a bit of a frustrating time trying to figure out why the
by_size subrouting was trying to compare the keys AND the values.
Sheesh. Otherwise, it's very straightforward. Switching from ascending
to descending was as easy as reversing the has subscripts in the by_size
sub.

/rs.