Answer for Ch. 14, Ex. 5 of Learning Perl Workbook.
Write a program to print a list of the files in the current directory,
sorted in order of their last modification line. Put the most recently
modified file at the beginning of the list.

I did this VERY differently than the workbook. (Well, I did the last one
very differently too, but we're talking about this exercise.) I
initially thought of using the -M flag to test it, and just switching it
into place of -s. But, that didn't show me the modification time in the
list, so I swapped it for a stat call. Then, once I had the sorted list,
I converted the epoch time to localtime in the printf.

It's not nearly as terse as the workbook's answer!

/rs.