Answer for Ch. 14, Ex. 5 of Learning Perl Workbook.
Create two arrays, one for the first names and one for the last names,
for the Flintstones characters such that the elements in each array
align with each other. Go through only one of the arrays with each and
use the array index to access the other array. For each character,
output their first and last names on the same line, with every
character’s full name on a separate line.

This question would have been better back in chapter 3, when they
introduced 'each'. I might have remembered it back then!

/rs.