Answer for Ch. 15, Ex. 2 of Learning Perl.
Write a program using given-when that takes a number as its input, then
prints “Fizz” if it is divisible by 3, “Bin” if it is divisible by 5,
and “Sausage” if it is divisible by 7. For a number like 15, it should
print “Fizz” and “Bin” since 15 is divisible by both 3 and 5. What’s the
first number for which your program prints “Fizz Bin Sausage”?

Pretty simple, again. Modulus FTW.

/rs.