Answer for Ch. 15, Ex. 2 of Learning Perl.
Modify the program from the previous exercise to also report if the
number is odd or even, if the number is prime (you find no divisors
other than 1 and the number itself), and if it is divisible by your
favorite number. Again, only use smart matching.

The exercise key for this one tests to see if 2 and my favourite number
(in this case, 17) are in the list of returned divisors from the
subroutine, and acts accordingly if they are. I chose to use separate
modulous calculations for both parity and divisiblity by 17. While my
solution is quite correct, and perhaps slightly more obvious to what is
happening, the text's answer is probably the more efficient one.

/rs.