Answer for Ch. 15, Ex. 1 of Learning Perl Workbook.
Use the smart match operator in a program that prompts the user for a
string and reports if that string is an element of an array you specify
in your program. You only have to match exact elements.

If I had used 'say' instead of 'print', you'd think I had copied the
solution instead of writing it myself. It was pretty obvious.

/rs.