Answer for Ch. 15, Ex. 2 of Learning Perl Workbook.
Modify your answer to Exercise 15.1 to prompt the user to enter several
names, one line per name. Report success if at least one of the names is
a key in a hash that you define in your program.

Minor modification from array to hash, and adding an array of elements
to test, rather than a single one.

/rs.