Answer for Ch. 15, Ex. 4 of Learning Perl Workbook
Write a program that prompts you for two lists of words. Use the smart
match operator to report if these two lists are the same or not. To make
things easier, enter each list on a single line then split that line on
whitespace.

The list with a leading space doesn't work, because it creates an empty
list item for the element, making the lists different. This could be
fixed by removing the leading whitespace.

/rs.