Answer for Ch. 15, Ex. 5 of Learning Perl Workbook
Use given-when and smart matching to implement a word guessing game. If
the user enters a string that exactly matches the secret word, report
success. If the string doesn’t match, try that string as a regular
expression pattern. If the user enters 'give up', stop the program.

I did this differently than the workbook, I given'd on the guess, rather
than the random word. I had to explicitly test the entered pattern as a
regex against the random word, just like in the workbook, but otherwise,
it worked fine.

/rs.