Answer for Ch. 16, Ex. 3 of Learning Perl.
Write a program to parse the output of the date command to determine the
current day of the week. If the day of the week is a weekday, print get
to work; otherwise, print go play.

The most challanging part of this exercise was figuring out the
appropriate flags for date to test weekends and weekdays.

/rs.