Chapter 16, Exercise 04
Write an infinite loop program that catches signals and reports which
signal it caught and how many times it has seen that signal before.
Exit if you catch the INT signal.

Wow. I really did this differently than the book. Here's how:

First, I used a generic callback function for all of the signals.

Second, I trapped a LOT more signals than the answer. I used a foreach
loop over an array to assign the callback function to each of the
entries in the SIG hash that I could find.

Third, I kept track of the signal counts in a global state hash,
instead of a single state variable for each. I kind of had to, I think,
since I was using a single sub for all of the signals. I was going to
make the hash a private state variable, but while reading found out that
it wouldn't work.

Fourth, I showed a summary table at the end of it all, using that
global hash. This was just because I like making tables of data, I
guess. It wasn't even technically part of the question requirement,
really.

All in all, I quite like my version. It's definitely longer, but I think
it's cooler, too.

/rs.