Chapter 16 Workbook Exercise 06
Modify your answer to Exercise 16.2 to execute the command using open
and a pipe. Read the input one line at a time through a filehandle and
report the same results.

Pretty simple conversion, I ended up switching to a while loop on the
suggestion of the answer key, as it did its magic as it was reading.
While it's not a big deal with a small data set, it would make the
program much easier to use if I were dealing with large amounts of data.

/rs.