Chapter 17 Exercise 03
Modify your answer to Exercise 2 to report the times using the
YYYY-MM-DD format. Use a map with localtime and a slice to turn the
epoch times into the date strings that you need.

I had a little trouble with this. I didn't really see the point of using
map to change the times, as it seemed like overkill. I also got confused
with the text's answer, with the seeming anonymous function in the
map... I ended up pulling the steps out into a function of their own
just so they made more sense to me.

I clearly need to figure maps a little more.

Also: all done the text! Now, only a handful of workbook exercises to
go.

/rs.