Chapter 17 Workbook Exercise 05
Given a single. multiline string stored in one scalar, transform it
so that each of the names at the beginning of each line have an
initial capital letter, and the rest of the name is in lowercase
letters.

This seems like something I would have done in Chapter 09, or something.

/rs.