Chapter 17 Workbook Exercise 07
Given the start of a program that looks up the characters’ bowling
scores, add a line to create the hash using a hash slice. The names
are the keys, and the values are their scores. Once you’ve done that,
change the program to create the hash using a map.

The first part was simple. The second part... I'm still not seeing how
that's more efficient. Perl maps seem like they need a lot more
attention than one page. I'm definitely going to need to come back to
these, thankfully in Intermediate Perl they touch on them again.

/rs.